Move an operation to pure
