+++ crates/swc_ecma_minifier/src/compress/pure/misc.rs
@@ -1,6 +1,7 @@
-use std::num::FpCategory;
+use std::{fmt::Write, num::FpCategory};
 
-use swc_common::{util::take::Take, DUMMY_SP};
+use swc_atoms::js_word;
+use swc_common::{iter::IdentifyLast, util::take::Take, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_utils::ident::IdentLike;
 
@@ -21,6 +22,176 @@ impl Pure<'_> {
                 self.remove_invalid(e);
             }
         }
+    }
+
+    pub(super) fn compress_array_join(&mut self, e: &mut Expr) {
+        let call = match e {
+            Expr::Call(e) => e,
+            _ => return,
+        };
+
+        let callee = match &mut call.callee {
+            Callee::Super(_) | Callee::Import(_) => return,
+            Callee::Expr(callee) => &mut **callee,
+        };
+
+        let separator = if call.args.is_empty() {
+            ",".into()
+        } else if call.args.len() == 1 {
+            if call.args[0].spread.is_some() {
+                return;
+            }
+
+            if is_pure_undefined(&call.args[0].expr) {
+                ",".into()
+            } else {
+                match &*call.args[0].expr {
+                    Expr::Lit(Lit::Str(s)) => s.value.clone(),
+                    Expr::Lit(Lit::Null(..)) => js_word!("null"),
+                    _ => return,
+                }
+            }
+        } else {
+            return;
+        };
+
+        let arr = match callee {
+            Expr::Member(MemberExpr {
+                obj,
+                prop: MemberProp::Ident(Ident { sym, .. }),
+                ..
+            }) if *sym == *"join" => {
+                if let Expr::Array(arr) = &mut **obj {
+                    arr
+                } else {
+                    return;
+                }
+            }
+            _ => return,
+        };
+
+        if arr.elems.iter().any(|elem| {
+            matches!(
+                elem,
+                Some(ExprOrSpread {
+                    spread: Some(..),
+                    ..
+                })
+            )
+        }) {
+            return;
+        }
+
+        let cannot_join_as_str_lit = arr
+            .elems
+            .iter()
+            .filter_map(|v| v.as_ref())
+            .any(|v| match &*v.expr {
+                e if is_pure_undefined(e) => false,
+                Expr::Lit(lit) => !matches!(lit, Lit::Str(..) | Lit::Num(..) | Lit::Null(..)),
+                _ => true,
+            });
+
+        if cannot_join_as_str_lit {
+            if !self.options.unsafe_passes {
+                return;
+            }
+
+            // TODO: Partial join
+
+            if arr
+                .elems
+                .iter()
+                .filter_map(|v| v.as_ref())
+                .any(|v| match &*v.expr {
+                    e if is_pure_undefined(e) => false,
+                    Expr::Lit(lit) => !matches!(lit, Lit::Str(..) | Lit::Num(..) | Lit::Null(..)),
+                    // This can change behavior if the value is `undefined` or `null`.
+                    Expr::Ident(..) => false,
+                    _ => true,
+                })
+            {
+                return;
+            }
+
+            let sep = Box::new(Expr::Lit(Lit::Str(Str {
+                span: DUMMY_SP,
+                raw: None,
+                value: separator,
+            })));
+            let mut res = Expr::Lit(Lit::Str(Str {
+                span: DUMMY_SP,
+                raw: None,
+                value: js_word!(""),
+            }));
+
+            fn add(to: &mut Expr, right: Box<Expr>) {
+                let lhs = to.take();
+                *to = Expr::Bin(BinExpr {
+                    span: DUMMY_SP,
+                    left: Box::new(lhs),
+                    op: op!(bin, "+"),
+                    right,
+                });
+            }
+
+            for (last, elem) in arr.elems.take().into_iter().identify_last() {
+                if let Some(ExprOrSpread { spread: None, expr }) = elem {
+                    match &*expr {
+                        e if is_pure_undefined(e) => {}
+                        Expr::Lit(Lit::Null(..)) => {}
+                        _ => {
+                            add(&mut res, expr);
+                        }
+                    }
+                }
+
+                if !last {
+                    add(&mut res, sep.clone());
+                }
+            }
+
+            *e = res;
+
+            return;
+        }
+
+        let mut res = String::new();
+        for (last, elem) in arr.elems.iter().identify_last() {
+            if let Some(elem) = elem {
+                debug_assert_eq!(elem.spread, None);
+
+                match &*elem.expr {
+                    Expr::Lit(Lit::Str(s)) => {
+                        res.push_str(&s.value);
+                    }
+                    Expr::Lit(Lit::Num(n)) => {
+                        write!(res, "{}", n.value).unwrap();
+                    }
+                    e if is_pure_undefined(e) => {}
+                    Expr::Lit(Lit::Null(..)) => {}
+                    _ => {
+                        unreachable!(
+                            "Expression {:#?} cannot be joined and it should be filtered out",
+                            elem.expr
+                        )
+                    }
+                }
+            }
+
+            if !last {
+                res.push_str(&separator);
+            }
+        }
+
+        report_change!("Compressing array.join()");
+
+        self.changed = true;
+        *e = Expr::Lit(Lit::Str(Str {
+            span: call.span,
+            raw: None,
+            value: res.into(),
+        }))
     }
 
     pub(super) fn drop_undefined_from_return_arg(&mut self, s: &mut ReturnStmt) {
